+++ tests/e2e/cases/runtime_eval/in.ts
@@ -1,2 +1,9 @@
 const source = "1 + 1";
 console.log(eval(source));
+
+function localEval(): unknown {
+    const localSource = "'local' + '-eval'";
+    return eval(localSource);
+}
+
+console.log(localEval());